Changing how multiple Enum/Typedef definitions are handled.
If multiple Enum definitions are not disjoint (i.e. one is a superset
of the other), then we will now return the superset. If the Enums are
disjoint, the definition will be removed completely from the dataset and
a warning will be emitted.

For Typedefs, if multiple definitions of different types are
encountered, a warning will be emitted and the definition will be
removed as well.